# backend/api.yaml
openapi: 3.0.3
info:
  title: Campus Hire
  version: 1.0.0
  description: APIs for Campus Hire.
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/company/registration:
    post:
      tags:
        - Company
      summary: Register a new company
      description: Creates a new company account with HR details.
      requestBody:
        required: true
      responses:
        "201":
          description: Company registered successfully
        "400":
          description: Invalid request
  /api/company/dashboard:
    get:
      tags:
        - Company
      summary: Company Dashboard Statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics returned successfully
  /api/company/create-drive:
    post:
      tags:
        - Company
      summary: Create Placement Drive
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        "201":
          description: Placement drive created
        "400":
          description: Invalid request
  /api/company/get-dashboard:
    get:
      tags:
        - Company
      summary: List Company Placement Drives
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of placement drives
  /api/company/update-profile:
    get:
      tags:
        - Company
      summary: Get Company Profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company profile returned

    put:
      tags:
        - Company
      summary: Update Company Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        "200":
          description: Profile updated successfully

    post:
      tags:
        - Company
      summary: Update Single Profile Field
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        "200":
          description: Field updated successfully
  /api/company/get-student/{id}:
    get:
      tags:
        - Company
      summary: Get Student Details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
        "404":
          description: Student not found
  /api/company/shortlist-student/{student_id}/{drive_id}:
    get:
      tags:
        - Company
      summary: Shortlist Student
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer

        - name: drive_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Student shortlisted
        "404":
          description: Application not found
...